Enable task deletion and normalize callback usage
Controller now defines callback functions that are supplied to ui
components (forms/buttons) and handle relevant ui state updates after
data is modified. For task deletion, data management is done directly in
controller so that card view is kept independent of data. Task form
submissions use custom event for this instead of callback although that
is currently only for testing the waters of that method (no good reason
for lack of uniformity)